interface/requirements.txt
nats-py
Jinja2
Flask
webcolors
pydantic
pytest